requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-generator==1.10
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
dnspython==2.6.1
email-validator==2.1.0.post1
fastapi==0.109.2
fastapi-login==1.9.3
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.25.2
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.24
langchain-openai==0.0.6
langfuse==2.16.0
langsmith==0.1.3
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.12.0
outcome==1.3.0.post0
packaging==23.2
pgvector==0.2.5
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.27
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
trio==0.21.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
wrapt==1.14.0
yarl==1.9.4